fixing read_only issues

+++ Heirustic.cpp
@@ -6,41 +6,7 @@
 
 float Heirustic::ComputeH(Node *N,int h_number) {
 
-    /*
-     *
-     * Heuristics description:
-     *
-     * Heuristic 1 - this heuristic calculates the number of blocks that are currently not in the correct 'position'.
-     * We do not count a block if it is currently in the arm of the crane.
-     *
-     * Heuristic 2 - this heuristic is similar to Heuristic 1. It calculates the difference between the current state
-     * and the goal state, but looks at the details of each block. If Block A in the goal state is supposed to be on top
-     * of Block B and under Block C and in the current state it is neither on top of B or under C, then we add 2 to the
-     * heuristic.
-     *
-     * Heuristic 3 - This heuristic adds 2 for every block that is not currently directly on top of the block on which
-     * it has to be in the goal state or if there is such a block somewhere below it (in the same pile).
-     *
-     * Heuristic 4 - this heuristic is twice the number of blocks that must be moved once plus four times the number of
-     * blocks that must be moved twice. A block that must be moved once is a block that is currently on a block different
-     * to the block upon which it rests in the goal state or a block that has such a block somewhere below it in the same
-     * pile. A block that must be moved twice is a block that is currently on the block upon which it must be placed in the
-     * goal state, but that block is a block that must be moved or if there exists a block that must be moved twice
-     * somewhere below it (in the same pile).
-     *
-     * Heuristic 5 - If blocks A and B are preventing each other from being moved to their goal positions, we call this
-     * a case of mutual prevention. This heuristic works the same way as heuristic 3, but adds 2 for each case of mutual
-     * prevention.
-     *
-     * Heuristic 6 - This heuristic is a combination of heuristics 4 and 5, except when there is a block that must be
-     * moved twice (heuristic 4) and is also one of the blocks in a mutual prevention (heuristic 5), we don't add anything
-     * for this mutual prevention.
-     *
-     *
-     */
-
     float penality = 0.0;
-
 
     if(h_number == 1)
     {
@@ -57,7 +23,6 @@ float Heirustic::ComputeH(Node *N,int h_number) {
         }
 
     }
-
 
     if(h_number == 2)
     {
@@ -87,15 +52,11 @@ float Heirustic::ComputeH(Node *N,int h_number) {
                 }
             }
         }
-
-
-
     }
 
 
     if( h_number == 3)
     {
-
 
         //Heuristic :  blocks below a block on the stack should have lesser value than the block
         for (int i = 0; i < N->n_stacks; i++) {
@@ -128,7 +89,7 @@ float Heirustic::ComputeH(Node *N,int h_number) {
 
         for (int i = 0; i < N->n_stacks; i++) {
 
-            int stack_size =   N->block[i].size();
+            int stack_size =   (int)N->block[i].size();
 
             if(i!=0) {
